Initial OS definition
This commit defines the starting point for our myHouse OS. We
define the os-core bundle via the myhouse-image-core recipe and
the myhouse-core class it inherits.

Two bundles are defined at this point, the beginnings of a first
product bundle are defined via the pacakgegroup-myhouse-speakers and
a developer bundle, to include tools for internal use when developing
products, has also been defined.

An additional image is available to build which provides the base
software load for the speakers product by generating an image with
the os-core and the speakers bundle already installed.

The README.md has been updated to demonstrate the goals of the layer.

Signed-off-by: Joshua Lock <joshua.g.lock@intel.com>